backend/requirements.txt
huggingface_hub
python-pptx
python-dotenv
requests
lxml
mcp
fastapi
uvicorn
